# api/console/resource/node_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.8.4
// - protoc             v5.29.1
// source: console/resource/node.proto

package resource

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationNodeCreate = "/api.console.resource.Node/Create"
const OperationNodeList = "/api.console.resource.Node/List"
const OperationNodeUpdate = "/api.console.resource.Node/Update"

type NodeHTTPServer interface {
	Create(context.Context, *CreateNodeRequest) (*CreateNodeReply, error)
	List(context.Context, *NodeListRequest) (*NodeListReply, error)
	Update(context.Context, *UpdateNodeRequest) (*UpdateNodeReply, error)
}

func RegisterNodeHTTPServer(s *http.Server, srv NodeHTTPServer) {
	r := s.Route("/")
	r.GET("/api/console/node", _Node_List3_HTTP_Handler(srv))
	r.POST("/api/console/node", _Node_Create3_HTTP_Handler(srv))
	r.PUT("/api/console/node/{id}", _Node_Update3_HTTP_Handler(srv))
}

func _Node_List3_HTTP_Handler(srv NodeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in NodeListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNodeList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.List(ctx, req.(*NodeListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*NodeListReply)
		return ctx.Result(200, reply)
	}
}

func _Node_Create3_HTTP_Handler(srv NodeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateNodeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNodeCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*CreateNodeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateNodeReply)
		return ctx.Result(200, reply.Data)
	}
}

func _Node_Update3_HTTP_Handler(srv NodeHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateNodeRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationNodeUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*UpdateNodeRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateNodeReply)
		return ctx.Result(200, reply.Data)
	}
}

type NodeHTTPClient interface {
	Create(ctx context.Context, req *CreateNodeRequest, opts ...http.CallOption) (rsp *CreateNodeReply, err error)
	List(ctx context.Context, req *NodeListRequest, opts ...http.CallOption) (rsp *NodeListReply, err error)
	Update(ctx context.Context, req *UpdateNodeRequest, opts ...http.CallOption) (rsp *UpdateNodeReply, err error)
}

type NodeHTTPClientImpl struct {
	cc *http.Client
}

func NewNodeHTTPClient(client *http.Client) NodeHTTPClient {
	return &NodeHTTPClientImpl{client}
}

func (c *NodeHTTPClientImpl) Create(ctx context.Context, in *CreateNodeRequest, opts ...http.CallOption) (*CreateNodeReply, error) {
	var out CreateNodeReply
	pattern := "/api/console/node"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationNodeCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out.Data, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *NodeHTTPClientImpl) List(ctx context.Context, in *NodeListRequest, opts ...http.CallOption) (*NodeListReply, error) {
	var out NodeListReply
	pattern := "/api/console/node"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationNodeList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *NodeHTTPClientImpl) Update(ctx context.Context, in *UpdateNodeRequest, opts ...http.CallOption) (*UpdateNodeReply, error) {
	var out UpdateNodeReply
	pattern := "/api/console/node/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationNodeUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out.Data, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}