# openapi.yaml
...
openapi: 3.0.3
info:
    title: Go-Kratos Console API
    description: Management console for Go-Kratos
    contact:
        name: Sendya
        url: https://github.com/sendya
        email: yladmxa@gmail.com
    license:
        name: Apache License
        url: https://github.com/google/gnostic/blob/master/LICENSE
    version: 1.0.0
paths:
...
    /api/console/namespace:
        get:
            tags:
                - Namespace
            operationId: Namespace_List
            parameters:
                - name: pagination.current
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.page_size
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.total
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.raw_total
                  in: query
                  schema:
                    type: string
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.NamespaceListReply'
        post:
            tags:
                - Namespace
            operationId: Namespace_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.console.resource.NamespaceCreateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.NamespaceCreateReply'
...
    /api/console/namespace/{id}:
        get:
            tags:
                - Namespace
            operationId: Namespace_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.NamespaceGetReply'
        put:
            tags:
                - Namespace
            operationId: Namespace_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.console.resource.NamespaceUpdateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.NamespaceUpdateReply'
        delete:
            tags:
                - Namespace
            operationId: Namespace_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.NamespaceDeleteReply'
    /api/console/node:
        get:
            tags:
                - Node
            operationId: Node_List
            parameters:
                - name: pagination.current
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.page_size
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.total
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.raw_total
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.NodeListReply'
        post:
            tags:
                - Node
            operationId: Node_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.console.resource.CreateNodeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.CreateNodeReply'
    /api/console/node/{id}:
        put:
            tags:
                - Node
            operationId: Node_Update
            parameters:
                - name: id
                  in: path
                  description: 节点ID
                  required: true
                  schema:
                    type: integer
                    format: int32
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.console.resource.UpdateNodeRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.UpdateNodeReply'
...
    /api/console/user/{id}/namespace:
        post:
            tags:
                - User
            operationId: User_BindNamespace
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.console.administration.BindNamespaceRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.administration.BindNamespaceReply'
    /api/console/user/{id}/namespace/{namespace_id}:
        delete:
            tags:
                - User
            operationId: User_UnbindNamespace
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
                - name: namespace_id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.administration.UnbindNamespaceReply'
...
components:
    schemas:
...
        api.console.administration.BindNamespaceReply:
            type: object
            properties: {}
        api.console.administration.BindNamespaceRequest:
            type: object
            properties:
                id:
                    type: string
                namespace_id:
                    type: string
...
        api.console.administration.UnbindNamespaceReply:
            type: object
            properties: {}
...
        api.console.resource.Agent:
            type: object
            properties:
                type:
                    type: integer
                    description: 部署状态
                    format: enum
                heartbeat_time:
                    type: string
                    description: 上次心跳时间
                version:
                    type: string
                    description: Agent版本
...
        api.console.resource.CreateNodeReply:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.console.resource.NodeInfo'
        api.console.resource.CreateNodeRequest:
            type: object
            properties:
                name:
                    type: string
                    description: 节点名称
                ip:
                    type: string
                    description: 节点IP
                auto_detect:
                    type: boolean
                    description: Agent自动上报开关
                zone_id:
                    type: integer
                    description: 可用区ID
                    format: int32
                env:
                    type: string
                    description: 环境变量
...
        api.console.resource.NamespaceCreateReply:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.console.resource.NamespaceInfo'
        api.console.resource.NamespaceCreateRequest:
            type: object
            properties:
                name:
                    type: string
                alias:
                    type: string
                description:
                    type: string
        api.console.resource.NamespaceDeleteReply:
            type: object
            properties: {}
        api.console.resource.NamespaceGetReply:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.console.resource.NamespaceInfo'
        api.console.resource.NamespaceInfo:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                alias:
                    type: string
                description:
                    type: string
                created_at:
                    type: string
                updated_at:
                    type: string
                running:
                    type: integer
                    format: int32
        api.console.resource.NamespaceListReply:
            type: object
            properties:
                pagination:
                    $ref: '#/components/schemas/protobuf.Pagination'
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.console.resource.NamespaceInfo'
...
        api.console.resource.NamespaceUpdateReply:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.console.resource.NamespaceInfo'
        api.console.resource.NamespaceUpdateRequest:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                alias:
                    type: string
                description:
                    type: string
        api.console.resource.NodeInfo:
            type: object
            properties:
                id:
                    type: integer
                    description: 节点ID
                    format: int32
                name:
                    type: string
                    description: 节点名称
                ip:
                    type: string
                    description: 节点IP
                agent:
                    allOf:
                        - $ref: '#/components/schemas/api.console.resource.Agent'
                    description: Agent状态
                created_at:
                    type: string
                    description: 节点创建时间
                updated_at:
                    type: string
                    description: 节点更新时间
                zone_id:
                    type: integer
                    description: 可用区ID
                    format: int32
                zone_code:
                    type: string
                    description: 可用区编码
                zone_name:
                    type: string
                    description: 可用区名称
                region_code:
                    type: string
                    description: 地区
                region_name:
                    type: string
                    description: 地区名称
                env:
                    type: string
                    description: 环境变量
                auto_detect:
                    type: boolean
                    description: 自动上报信息
        api.console.resource.NodeListReply:
            type: object
            properties:
                pagination:
                    $ref: '#/components/schemas/protobuf.Pagination'
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.console.resource.NodeInfo'
...
        api.console.resource.UpdateNodeReply:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.console.resource.NodeInfo'
        api.console.resource.UpdateNodeRequest:
            type: object
            properties:
                name:
                    type: string
                    description: 节点名称
                ip:
                    type: string
                    description: 节点IP
                auto_detect:
                    type: boolean
                    description: Agent自动上报开关
                zone_id:
                    type: integer
                    description: 可用区ID
                    format: int32
                env:
                    type: string
                    description: 环境变量
                id:
                    type: integer
                    description: 节点ID
                    format: int32
...
        protobuf.Pagination:
            type: object
            properties:
                current:
                    type: integer
                    format: int32
                page_size:
                    type: integer
                    format: int32
                total:
                    type: integer
                    format: int32
                raw_total:
                    type: string
            description: 通用的分页信息
    securitySchemes:
        ApiKeyAuth:
            type: apiKey
            name: Authorization
            in: header
...